LayoutTests/platform/blackberry/fast/filesystem/file-writer-gc-blob-expected.txt

CONSOLE MESSAGE: line 163: ReferenceError: Can't find variable: webkitRequestFileSystem
FAIL: Timed out waiting for notifyDone to be called
Test that a blob won't get garbage-collected while being written out by a FileWriter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
